Add Timer for Requests to Support Business CDN Evaluation
Edgio, the company running CDNs that power the .NET ecosystem, has declared bankruptcy. We are conduction an analysis of CDNs across the ecosystem because of this. Providers such as Verizon have also held our content. Timeouts are a significant issue and while much better than before still represent a concern for us, particularly in Asia.

This PR enables data collection for these web requests.
Note that this is not the most unbiased data. I couldn't find a place in the get requests where we could actually get a timer for just the web requests, without dealing with axios. That's likely possible -- chrome seems to be able to do it. (Inspect element, network, click on something and view timing. They also have a timing API. But their toolkit is much more expansive than ours.) One example, is that So far on my testing, the download of the installer itself is faster than the releases.json. Which is kinda crazy if true; The installer.exe download took about 1/5th the time of the releases json request the first time, and 1/2 the time the 2nd time. I would believe that's due to some confounding variable -- axios casts the response type to a json object if it can, so that cast cost is likely included in the releases.json but not in the installer download.

This data seems sufficient, for now. Our main goal is to look at network location and time as well as CDN perf at large.