Remove batch true
This is so multiple commits can run PRs at the same time. This has been turned on for a long time and I'm not sure why it was.

Thank you @miyanni for helping figure this out, it was a weird an unexpected thing to dig through from the initial problem set.