Add CG Manually to Configure Exclude Dir
Hopefully this will override the 1ES auto injected task because I can't find a way to configure that one. I've also asked them if that's possible.

The .vscode test package.json describes the built-in handlebars extension, not the handlebars npm package that has the vulnerability, so it should be excluded.

CG is auto injected in each task so I had to try adding it to each one.