Improve Perf by Conditioning Taskkill + Fix whitespace and other code style no's (#2088)
* Fix whitespace and other code style no's

Please see https://github.com/dotnet/vscode-dotnet-runtime/issues/2082

I'll use this as an opportunity to introduce a needed perf improvement per a request from another team as well since we are in the business of improving the code.

taskkill is currently called all of the time. I added that task kill only to kill timed out installer processes (https://github.com/dotnet/vscode-dotnet-runtime/pull/1859/commits/43f7494777f3d757134801f1f268735c9dd334a6#r1669397622), but the code seems to run at times when I wouldn't expect it to. It likely doesn't need to be run at all, I can put a condition on it.

* Only perform task kill when given a custom option to.

Forwarding this to the process runner should be fine - it accepts arbitrary args and this is a fast way to send more args without expanding the interface.

* make linter happy

* Fix formatting

* Remove task kill

Also fixes whitespace