Support other possible future domains for the installer files.
Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/2126
Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/2125

Workaround fix for https://github.com/dotnet/vscode-dotnet-runtime/issues/2123

We have a list of possible safe domains as part of our threat model which is needed to verify the source of truth when we run executables with elevated permission. In the past releases json only hosted downloads on `download.visualstudio.microsoft.com` but now it can be `builds.dotnet.microsoft.com`. I've added some more urls to our azure front door and other CDNs in the event that we decide to change to those.

The long term fix for this would be if signatures are ever published for releases.json to verify those instead, but that is not ready yet and this change needs to go in.