Improve Null Handling
I noticed in the telemetry some number of errors due to trying to access content-headers when headers was null. I don't know why axios would return a response but with undefined headers: https://github.com/axios/axios/issues/4812 based on this it may be a bug of some sort. Anyway, we want to handle that and I didn't know you could chain ?.['index'].

At the same time, many calls checked if something !== null. null is falsey, as is undefined, as is calling foo?.something when foo or something is undefined. We don't need all of those. Now there is some logic where the difference between null and undefined matter and I didnt touch those. Or cases where it easier to read but those are far and few between.