Handle x?.length === 0 better
Sometimes things can end up being undefined even if their typescript type is only string and not string | undefined. This was demonstrated in the above commit where axios object headers could be undefined. It's certainly possible to circumvent the axios type system. We do handle some objects with any types due to the dynamic nature of objects which could change form (I dont think this was the best decision but it is what it is, and I don't see us prioritizing a refactor.) Some of these types are safe but to be more cautious and or make the compiler happy, I've done this.

If you execute string?.length === 0, if the string is undefined, undefined === 0 is not true. Generally in a check like this I would want undefined lengths to be treated as 0. This is pretty much safe guarding our code somewhat to handle this.