Use a random distribution for cache reporting
So, the information we have about caching is useful to see if it's helpful or not and giving a perf boost/causing issues. But also, we are sending like 500k of them per hour and the data team is not very enthused about this. We only need a small random sample to get what we need so this is a better use of $

suppressTelemetry is checked in the TelemetryObserver to not send it.