Improve Performance by Utilizing Shells Properly
There are many ways to execute commands in nodejs via `child_process`.
Unfortunately, we were using `spawnSync` which blocks the entire thread (all of vscode extensions share one thread) which is highly inefficient.

I learned MANY things by reading various docs and such about all of the ways to do this to figure out the 'best' thing to do for our situation.

Here are some other fun improvments:

- You can spawn with powershell.exe shell! Running it with cmd (default shell) spawns a shell which spawns a shell, which is a huge unnecessary perf hit.

- You can go shell-less when running direct file executables to improve performance.

What more we could do (this is already a huge boost):
- The WSL check calls sync methods still because it needs to be called in the activate function to disable commands from showing up in WSL that are not supported. This makes startup slow. I would rather just have worse UX for WSL by allowing the command to appear but saying it doesn't work so it doesn't make vscode way slower for everybody.

Try to find places where executables can be used to reduce shell spawns.

# Ways to Run Commands

exec - spawns a shell. returns callback with error, stderr, stdout after waiting for it to complete. Any exit code other than 0 will be an error. error.signal for when a signal terminated it.

util.promisify(function) turns a callback into await or thenable (same thing) object with stderr or stdout and rejects with error with same object but with an error instead with exec.

execFile - same as exec but shellless - must be a file to exec instead of any generic command. No shell spawn by default thus more efficient as you should give it the path to the executable instead of expecting it to be found by the shell. Utilize when possible

spawn - shelless by default (which is faster if you dont need a shell but both have this option) spawns a new process with open stream communication for the command (shell or no shell) and streams data live. meant for 'interactive' commands and is better for huge data streams as you can chunk it.

fork - spawn a new process that can communicate with the parent, no shell support

spawnSync - Blocks the Entire Node Thread until its done. Same with execSync.

shelljs - a library for running native unix commands cross-plat

The ways to access the 'signal' or 'status code' are different and while spawn provides an easier interface to get to them in theory, if you read the docs on the callback return object possibilities from node you will see its applicable in either situation.

certain commands are implemented in nodejs such as chmod

# more reading
https://nodejs.org/api/child_process.html#child_processexeccommand-options-callback

https://stackoverflow.com/questions/14332721/node-js-spawn-child-process-and-get-terminal-output-live
https://stackoverflow.com/questions/48698234/node-js-spawn-vs-execute