Remove Property for checking if WSL is supported
This must run during activation and before a certain point of VS Code Startup, however,  you can't make this async. This will make startup slower and the point was to hide unsupported commands on WSL so users didn't have the struggle of seeing it was unsupported.

However, it's not worth slowing everyone down for this so now it just happens when you try the command.