Code Cleanup: Improve Null Handling (#2134)
* Improve Null Handling

I noticed in the telemetry some number of errors due to trying to access content-headers when headers was null. I don't know why axios would return a response but with undefined headers: https://github.com/axios/axios/issues/4812 based on this it may be a bug of some sort. Anyway, we want to handle that and I didn't know you could chain ?.['index'].

At the same time, many calls checked if something !== null. null is falsey, as is undefined, as is calling foo?.something when foo or something is undefined. We don't need all of those. Now there is some logic where the difference between null and undefined matter and I didnt touch those. Or cases where it easier to read but those are far and few between.

* Handle x?.length === 0 better

Sometimes things can end up being undefined even if their typescript type is only string and not string | undefined. This was demonstrated in the above commit where axios object headers could be undefined. It's certainly possible to circumvent the axios type system. We do handle some objects with any types due to the dynamic nature of objects which could change form (I dont think this was the best decision but it is what it is, and I don't see us prioritizing a refactor.) Some of these types are safe but to be more cautious and or make the compiler happy, I've done this.

If you execute string?.length === 0, if the string is undefined, undefined === 0 is not true. Generally in a check like this I would want undefined lengths to be treated as 0. This is pretty much safe guarding our code somewhat to handle this.

* Improve readability

* Fix lint

* Fix bug with null handling
0 is falsey so the 'improvement' was wrong

* fix bug with type of telemetry:

* denoate that  disableextensions is not enabled for local testing

* looks like you can but it may cause some weird behavior

* fix newline

* fix nl