Fix bug getting the status of a command
The nodejs spec for the special cased promisify version of exec mentions that it throws an object with an error, stdout and stderr. It looks like the object it throws is not an error in an object but rather an error object. So calling .error is incorrect.

This fixes the handling of cancellations.