Cache URL Requests Better
The old web request library was a class which each owned their own axios instance.
This used to work when the cache storage was persistent and stored within VS Code. But now that its a local memory cache, I migrated the class to a singleton so the cache actually gets used by everyone.
For why its not persistent storage, there is a bug in VSCode described here which makes doing that not work well: https://github.com/dotnet/vscode-dotnet-runtime/pull/1540

That was a very subtle bug and rare so I dont want to deal with figuring out if it's better or not in a hot fix of releases, but I figured this was simple enough that it would be ok to fit it in. (Maybe I was a bit optimistic.)

Thank you to @baronfel as well for doing some testing around with the axios cache library.