Fix timer for downloading files
The file download response type is a 'stream' which means the response gets sent back with an open stream to pipe the file.
This means the recorded end time for the webrequest was for the first few chunks of data and the initial stream pipe handover.
Now it waits to record the data until after that is done.

To simplify the logic, it still records every web request but recognizes if its in a stream mode not to record that data and let the end of the stream handler callback report it.