Dont use Cleanup Interval as it Causes Hang Forever
CleanUp Interval spawns a callback which never finishes probably because it was designed for use in a web browser so when you close it, all of the js process is killed.
This causes our tests to never finish or any code relying on it never exits.

I implemented a different method to clean it up -- with a destroy () method (sadly no easy way to add RAII in JS) for tests.
Otherwise the running function object will realize when its been deleted as its been bound with the (this) scope.

You can bind automatically in es6 using an arrow function but that doesn't seem to bind the correct this ->
there is also the that = this convention with .bind(this) but the simplest approach to me is to bind it by making it part of the scope (aka a class member.)