Fix 2.2.9 Bugs for Release. Command Status, Web Cache, & Web Timing (#2169)
* Fix bug getting the status of a command

The nodejs spec for the special cased promisify version of exec mentions that it throws an object with an error, stdout and stderr. It looks like the object it throws is not an error in an object but rather an error object. So calling .error is incorrect.

This fixes the handling of cancellations.

* Dont retry download if it was already downloaded

This also fixes a callback which was added incorrectly during the async refactor as it did not wait on the callback.

* Cache URL Requests Better

The old web request library was a class which each owned their own axios instance.
This used to work when the cache storage was persistent and stored within VS Code. But now that its a local memory cache, I migrated the class to a singleton so the cache actually gets used by everyone.
For why its not persistent storage, there is a bug in VSCode described here which makes doing that not work well: https://github.com/dotnet/vscode-dotnet-runtime/pull/1540

That was a very subtle bug and rare so I dont want to deal with figuring out if it's better or not in a hot fix of releases, but I figured this was simple enough that it would be ok to fit it in. (Maybe I was a bit optimistic.)

Thank you to @baronfel as well for doing some testing around with the axios cache library.

* Fix timer for downloading files

The file download response type is a 'stream' which means the response gets sent back with an open stream to pipe the file.
This means the recorded end time for the webrequest was for the first few chunks of data and the initial stream pipe handover.
Now it waits to record the data until after that is done.

To simplify the logic, it still records every web request but recognizes if its in a stream mode not to record that data and let the end of the stream handler callback report it.

* fix changelog

* fix interfaces to be aligned for mocks

* fix interfaces for mocks

* fix lint

* fix test

* Dont use Cleanup Interval as it Causes Hang Forever

CleanUp Interval spawns a callback which never finishes probably because it was designed for use in a web browser so when you close it, all of the js process is killed.
This causes our tests to never finish or any code relying on it never exits.

I implemented a different method to clean it up -- with a destroy () method (sadly no easy way to add RAII in JS) for tests.
Otherwise the running function object will realize when its been deleted as its been bound with the (this) scope.

You can bind automatically in es6 using an arrow function but that doesn't seem to bind the correct this ->
there is also the that = this convention with .bind(this) but the simplest approach to me is to bind it by making it part of the scope (aka a class member.)

* fix lint

* copilot fixes

* code cleanup

* Always create a Package to test

* Fix Typo

Co-authored-by: Jacques Eloff <joeloff@users.noreply.github.com>

* Dont consider no dotnet install as failure on lnx

---------

Co-authored-by: Jacques Eloff <joeloff@users.noreply.github.com>