Count the lock as stale at timeout - retryTime *2
The lock file being stale happens when a process dies and never returns/exists. If this is the same as the timeout duration, then the staleness check will never work (e.g. if another process holding it got stale and we check/set that only once we've timed out, it will possibly not recover properly.)

retryTime is the estimated amount of ms it takes to retry. Some time may pass when the program is executing other code so it is not exact. If it goes over, we may hit the timeout. So, we multiply it by 2 to be safe.