Handle VS Code Termination
We give vscode & at the end of the command to run the command in the BG and return 0 once it starts the shell script.
But their lib encases the command in a " " and echos it to bin bash. It doesnt look like the & is completely effective but I do believe it serves some purpose to run it in the bg which I forget.

Since we cant await for that thing to finish, we can instead let it go and then run our check waiting for the ok.txt file. But we dont want the control flow to continue if the command spawn fails. So we can have the vscode lib error written to a singleton which is accessible from any control flow state. Then we can check for the error or ok.txt file, that way if theres an error we can immediately stop control flow.