Hold Lock Over Machine State Changing Operations (Sync Code Is Now Async) (#2177)
* Generalize Function to Execute Code Holding a Lock

* wrap the local install in a folder based lock

* Require a lock file path instead of relying in file local directory

* Create install dir if it DNE

* Wrap Global State Logic in a Lock as well

This is because only 1 installer can run at a time (same with holding the apt lock or dnf lock etc)

* Prevent circular import

* Use different timeout values and check lock values

* Dont count lock timeout as installer problem

* Fix lint

* Fix timeout time calculation

Its not a time out max for total time but for retry https://www.npmjs.com/package/proper-lockfile

* It doesn't release the lock with an exception with the latest change, see if this is the fix or I need to call .unlock()

* Move catch logic into 2 thoughts - 1. lock errors 2. function to run errors

* improve comment

* Linux output should be executable

* hold the lock when updating dotnet on lnx

* fix lint

* fix lint

* Remove other lock oriented code so its all shared

* Move sudo spawn check to shared location... a bit of bloat, but better than a global variable.

* handle error properly

* shorten time to check for sudo proc dying

* Hande PKill Gracefully

* fix code

* unlock manualy

* always reject because if we dont spawn a sudo process we cant operate properly

* the error was meant to be used from the lib

* wait for proc to end because it wont end the other proc

* Consider the error from vscode may not have a name property

* try to fix callback

* fix logic

* Utilize Throw over Rejection

Currently its just getting ignored as a rejected promise instead of bubbling up the callstack

* Use resolve, reject

* add back in necessary logic

* Await the unlock

.then() will be awaited by the await call. Catch also is just a function that rejects or returns a resolved promise. If we dont await the unlock failure, we may throw the lock error at an unexpected time, which could cause any number of bad interactions or concurrency bugs

* Count the lock as stale at timeout - retryTime *2

The lock file being stale happens when a process dies and never returns/exists. If this is the same as the timeout duration, then the staleness check will never work (e.g. if another process holding it got stale and we check/set that only once we've timed out, it will possibly not recover properly.)

retryTime is the estimated amount of ms it takes to retry. Some time may pass when the program is executing other code so it is not exact. If it goes over, we may hit the timeout. So, we multiply it by 2 to be safe.

* Move variable calc down for perf

We dont use those until later, probably the compiler optimizes it anyways but worth trying

* End the sudo process upon failure

* Wait Time should be in Ms

* Update variable name for clarity

* await the exit otherwise it may exit while the other one is running

* Wait for only some of the timeout time to end the sudo process

* Dont try to kill the sudo process if it never spawned

this takes  a lot of waiting time and causes other things to timeout

* Check if there is an existing sudo process from code

* Fix Lock Timer

* Check if Sudo Spawn was successful instead of checking if we asked for the lock

- Removes isLive check as the function to spawn it should now reject if it fails anyways and this just causes perf regression

- Check if sudo spawned successfully instead of if we got the lock because if not we don't want to wait very long

- Dont wait for the sudo process for the same duration we are waiting for the lock as this will cause the lock error to possibly occur before the sudo one

* Set a timeout in case vscode sudo library never exits

* Remove unnecessary syntax sugar and fix timeout check

* Record response time var in log

* Fix timeout calc

* Don't await vscode spawn as it does live forever, actually

* Handle VS Code Termination

We give vscode & at the end of the command to run the command in the BG and return 0 once it starts the shell script.
But their lib encases the command in a " " and echos it to bin bash. It doesnt look like the & is completely effective but I do believe it serves some purpose to run it in the bg which I forget.

Since we cant await for that thing to finish, we can instead let it go and then run our check waiting for the ok.txt file. But we dont want the control flow to continue if the command spawn fails. So we can have the vscode lib error written to a singleton which is accessible from any control flow state. Then we can check for the error or ok.txt file, that way if theres an error we can immediately stop control flow.

* return all of it in the lock

* Hold the lock while running a cmd after checking if sudo master proc is live

This is a hack but testing this for now

* dont need to unlock manually

* extend initial check delay as it may miss it within 250 ms

* fix linting issues and make linter stricter:

* fix lint

* hold the lock during the entire sudo process, including spawning it. The vscode async logic is in a catch() block and it is async so it should not return up the code flow and cause weird unlocking issues, its errors are swallowed and propagated

* Hold lock during entire install check process

* fix lint

* dont hold lock during update

* Log lock information better

* Check Lock Ownership before delete

* the lock is a dir

* Clean up dpkg lock

If pkilled then we may hold it

* resolve lock key path so relative path doesnt cause issue

* Handle when it seems lib will try release even if 'acquired' code never ran?

idk why this doesnt bubble to the catch but iut becomes a dotnetcommandfailed after the releasedevent and never goes further...

actually looks like it might be an error event but anyways we can ignore that.

* Ignore not owned by you

* Only ask for invalid integrity if finished downloading

* Fix validation for linux

We changed it so the executable is returned and not the dotnet folder, this tells it that the folder is not what is returned so to check the file as a file and not a directory

* Fix typo

Co-authored-by: Jacques Eloff <joeloff@users.noreply.github.com>

* Dont cache exe downloads and pkg downloads as the cache lib doesnt work with downloads

* Fix bug where uninstalling externally on linux after installing in vscode would fail to install again later

This is because the validator code was not correct

* Fix existing logic

* Update change list for 2.3.0 since this has breaking changes

---------

Co-authored-by: Jacques Eloff <joeloff@users.noreply.github.com>