Don't record all of the environment in the log
We only need certain information to help debug logs, and when customers submit them to us, they are less likely to redact everything. It's been helpful having this information for debugging and this will make it harder to debug in case there are variables that I didn't think of, that a user might set that'd cause a bug, but for the sake of collecting less information on the log file we should do this.

This also hashes the string of the env used as the key to cache commands, since each command environment may result in different results, we want to still have a 'unique' enough ID for it without storing this data in our cache.