Update Packages - Important one : Axios-Cache-Interceptor
The version of Axios we upgraded to caused issues in https://github.com/arthurfiorette/axios-cache-interceptor/issues/953.

We weren't using the Cache before, but now that we are, it seems to have caused some issues give 1.6.0 of the cache interceptor is not compatible with the version of axios we were using.