Skip Parsing Empty Dotnet Executable
Callers may try to find --list-runtimes with an empty string, especially if the existingPath setting is empty, which can return '""' instead of ''. This is a performance improvement.

In addition, it looks like a recent change I did break some of the logic to find a working command. options?.at() is not valid when options is not provided as an array because it is  an object type. [] indexing does the same thing and is safer, as indexing on an object will just return undefined, unless it has a '0' key, but that is never expected here.