Improve Cache Time Duration
We did some testing and performance analysis and would like to include it in this release, that most of the time on average it takes about 3 seconds before this is called again. It is an edge case where someone installs dotnet externally in the middle of loading the process server, and cdk and c# will not refresh by calling findPath again, so they will be broken in this case regardless of if the cache is correct or not.