Fix Proxy Lookup
The old lookup was as follows:
`ctx?.proxyUrl ?? configuredProxy`.

But `??` does not evaluate truthfulness. It evaluates nullness.
The default empty value option for getting the string setting is '' or '""'.
Setting this to the proxy agent will make it so it never finds the user setting and it will also make it fail: https://github.com/search?q=repo%3ATooTallNate%2Fproxy-agents%20invalid%20url&type=code

And it wasn't in the try block, as this is a best effort approach it should be.