Fix Linux Tests By Centralizing Distro Version Info
The tests are conditioned to check the support status of dotnet, etc.
Last night the CI runners got updated to Ubuntu 24.04. 24.04 does not have .NET 9 support built-in, which means the versions of dotnet either went up or down.

The old logic was somewhat spread out.
This is still not ideal but this at least reduces the points of contact that need updated.
Maybe this is a bit too extensive of code cleanup for fixing a breaking pipeline, though I'm the only one who will be blocked by this for now, so I think it's ok.