Implement Node Mutex (#2217)
* Implement Node Mutex

# Summary

Introduced `NodeIPCMutex`.

This is a class that provides a mutex (mutual exclusion) lock using IPC (Inter-Process Communication) on Node.js.
 * It will work across multiple processes and async code in the same process.
 * It uses a named pipe on Windows and a file descriptor on Linux and OS X to create the lock.

# Why are you writing all of this when it seems like this should already exist somewhere out there in node.js land?
### Hint: It Doesn't ( :( )

 * Many locking mechanisms such as lockfile and proper-lockfile run into issues when processes are killed or die unexpectedly, because they rely on the file system.
 * windows-mutex is another library, which is a wrapper around a C mutex, however it is Windows only.
 * async-mutex is another node.js library, but it stores a list of owners in memory, which may cause issues if multiple processes try to acquire the same lock.
 * Named pipes on windows die when processes die. There is no equivalent check on Linux or OS X, but IPC allows us to check whether the process is still alive or not so we don't deadlock

# How can I be reassured this wont be potentially catastrophic / have unexpected corner cases

 * Much of this code is inspired by VSCode's IPC code, which is a well tested approach. However, their approach focuses on permitting only one main VS Code Process to run at a time.
 * https://github.com/microsoft/vscode/blob/main/extensions/git/src/ipc/ipcServer.ts#L15
 * https://github.com/microsoft/vscode/blob/main/src/vs/code/electron-main/main.ts#L318

As much as I would love to have a backup option of the old code, we cannot have some code trying a backup option if it does not work, because then everything will be broken if 2 different locking mechanisms are utilized. It could be an environment flag, but I don't trust people will remember for all of the instances of vscode to set it correctly.

# Changes:

-> Rewrote the code to imply call around this wrapper without having to change any other logic.

-> The VS Code, (code) is well tested as it is run by all of the vscode instances. There is obviously an inherent risk that this may cause unexpected issues, but with some thorough testing and backup from the existence that this previous logic is good, I hope that we can trust this code.

-> We could probably remove some of the logging events/change the interface, but actually, I am trying to simply just change this without having to touch anything else.

* Don't consider socket.on(error) as stale, & Code Cleanup

- Consider that the lock is not necessarily stale if we don't get econnrefused.
- Code cleanup and linter response

* Remove proper-lockfile dependency

We don't need that anymore.

* Remove now unused code

* Don't try to give a filepath as a lock obj

* Only send lock success data 1 % of the time

So we don't blast the data team with a huge amount of data

* Add some tests

code is still kinda wip but it's also test code

* Set a timeout for all mutex tests

* Simplify Connection Failure Logic

* Log error when unable to remove stale lock

* ensure release is always called

* Do we need to destroy the socket on the err case on mac?

It seems EADDRESSINUSE is happening on Mac, such that the FD is dead but the listener never stops, then we have 2 listeners listening on an empty thing which makes it fail in many ways.

Not sure if this will help or not yet. May revert.

* Remove Chmod

this makes everything slower

- if you run vscode under sudo you are already so doomed and it is required to use a different user folder so they should not conflict
- windows doesnt need to delete an admin pipe as pipes always rm when processes die
- vscode doesnt chmod and had no problems

* Ensure the socket is always destroyed

I still don't know why mac is not able to reconnect.
Is it also because the task runner doesn't self die for some reason?

* Report test failure cause on another line

* Always increment retries

* Consider that multiple callers might try to delete the lock at the same time

We can't rm if 2 callers delete it, so this seems like a bug to me.

* Add a delay if we try to clean up multiple times

this seems to be happening on Mac. but we dont want to slow down the first base case.

* Wrap the entire server creation in a finally

It seems on MacOS, the server doesn't get released in the error case, which causes ECONNREFUSED, and the FD is dead, but there are 2 servers that are alive :thinking:

* Revert "Wrap the entire server creation in a finally"

This reverts commit 9ecd73ef18b006f71f6ce79d0bba12e4df17075f.

* Try to force release

* Improve logging and add more delay for tests

Maybe this will help understand why Mac is failing. actionId could probably be part of the class but idk if I'll keep it.

* Improve Retry Logic

The message was not accurate.
Also removes the extra release on mac since it didn't seem to help any.

* Add a delay factor to tests

* Consider that tests may retry and then cause the lock to timeout on other tests

Also, the file path of locks was too long for mac which caused issues.

* Shorten length of locks

* Add opt out environment variable

in case /temp/ is too long for some reason and the user cant change that or doesnt want to

* Simplify Error

* Fix typo

Co-authored-by: Michael Yanni <MiYanni@microsoft.com>

* Fix typo

Co-authored-by: Michael Yanni <MiYanni@microsoft.com>

* Fix typo

Co-authored-by: Michael Yanni <MiYanni@microsoft.com>

* Consider FreeBSD and OSX 10.9 Length Limit

* Extract error message so it is customizeable.

This also removes the old '.lock' from the string since that is not relevant anymore.

* Improve documentation

---------

Co-authored-by: Michael Yanni <MiYanni@microsoft.com>